+++ app/Modules/Comunicacao/UI/Livewire/Template/TemplateManager.php
@@ -24,7 +24,9 @@ class TemplateManager extends Component
 
     public function mount()
     {
-        abort_if(!auth()->user()->hasRole('dev|admin'), 403);
+        abort_if(!feature('template.listar'), 403, 'Módulo desativado.');
+        abort_if(!auth()->user()->hasRole('dev') && !auth()->user()->can('template.listar'), 403);
+
         $this->breadcrumbs = BreadcrumbHelper::generate();
         $this->permiteGrid = false;
     }
@@ -41,6 +43,9 @@ class TemplateManager extends Component
 
     public function excluir($id)
     {
+        abort_if(!feature('template.excluir'), 403);
+        abort_if(!auth()->user()->hasRole('dev') && !auth()->user()->can('template.excluir'), 403);
+        
         try {
             EmailTemplate::findOrFail($id)->delete();
             $this->dispatch('sucesso', msg: 'Template removido com sucesso!');
